# eventify/ eventify-template/eventify-backend/src/main/java/com/eventify/eventify_backend/api/dto/response/AuthResponse.java
package com.eventify.eventify_backend.api.dto.response;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class AuthResponse {


    private String token;
    private String role;
}